+++ memory/environment.go
@@ -10,21 +10,26 @@ type Environment struct{
 	Values map[string]interface{}
 }
 
-// todo throw runtime error
+// 赋值语句
 func (e *Environment) Assign(name lexer.Token, value interface{}) {
 	if _, OK := e.Values[name.Lexeme]; OK {
 		e.Values[name.Lexeme] = value
+		return
+	}
+	if e.Enclosing != nil {
+		e.Enclosing.Assign(name, value)
 		return
 	}
 	err := Err.RunTimeError{Msg:"Undefined variable '" + name.Lexeme + "'.", Token: name}
 	panic(err.Error())
 }
 
+// 定义变量
 func (e *Environment) Define(name string, value interface{}) {
 	e.Values[name] = value
 }
 
-// todo throw runtimeError
+// 获取变量的值
 func  (e *Environment) Get(name lexer.Token) interface{} {
 	if data, OK := e.Values[name.Lexeme]; OK {
 		return data