+++ CHANGELOG.md
@@ -7,9 +7,13 @@ The format is based on Keep a Changelog, and this project adheres to Semantic Ve
 ### Added
 - tons of shit
 
+## 0.3.0 - 2019-03-17
+### Added
+- add scope for global and local variable
+
 ## 0.2.0 - 2019-03-07
 ### Added
-- parser to evaluate AST statements
+- parser to execute AST statements
 - add variable
 
 ## 0.1.0 - 2019-03-07